# app/lib/api/swagger/v1.yaml
swagger: '2.0'
info:
  title: SnitchDNS API
  version: "1.0.0"
schemes:
  - https
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-SnitchDNS-Auth
basePath: /api/v1
produces:
  - application/json
consumes:
  - application/json
security:
  - APIKeyHeader: []
paths:
  /zones:
    get:
      summary: Return all domains the user has access to
      tags:
        - zone
      parameters:
        - name: page
          in: query
          required: false
          type: integer
        - name: per_page
          in: query
          required: false
          type: integer
        - name: search
          in: query
          required: false
          type: string
          description: Domains to search for
        - name: tags
          in: query
          required: false
          type: string
          description: Comma separated tags to search for
      responses:
        200:
          description: Will return an array with all zones the user can access
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/Pagination'
              - properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/Zone'
        401:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
    post:
      summary: Create a new zone
      tags:
        - zone
      parameters:
        - in: body
          name: zone
          schema:
            type: object
            required:
              - domain
              - active
              - catch_all
              - forwarding
              - regex
              - master
              - tags
            properties:
              domain:
                type: string
              active:
                type: boolean
              catch_all:
                type: boolean
              forwarding:
                type: boolean
              regex:
                type: boolean
              master:
                type: boolean
              tags:
                type: string
      responses:
        200:
          $ref: '#/definitions/Zone'
        401:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
  /zones/{zone}:
    get:
      summary: Retrieve a specific zone
      parameters:
        - name: zone
          description: If this parameter is a number it will be treated as the Zone Id, otherwise it will be treated as the domain
          in: path
          required: true
          type: string
      tags:
        - zone
      responses:
        200:
          $ref: '#/definitions/Zone'
        401:
          $ref: '#/definitions/Response'
        404:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
    post:
      summary: Update a zone
      tags:
        - zone
      parameters:
        - name: zone
          description: If this parameter is a number it will be treated as the Zone Id, otherwise it will be treated as the domain
          in: path
          required: true
          type: string
        - name: data
          description: Zone data to update
          in: body
          schema:
            type: object
            properties:
              domain:
                type: string
              active:
                type: boolean
              catch_all:
                type: boolean
              forwarding:
                type: boolean
              regex:
                type: boolean
              tags:
                type: string
      responses:
        200:
          $ref: '#/definitions/Zone'
        401:
          $ref: '#/definitions/Response'
        404:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
    delete:
      summary: Delete a zone
      tags:
        - zone
      parameters:
        - name: zone
          description: If this parameter is a number it will be treated as the Zone Id, otherwise it will be treated as the domain
          in: path
          required: true
          type: string
      responses:
        200:
          $ref: '#/definitions/Response'
        401:
          $ref: '#/definitions/Response'
        404:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
...
  /search:
    get:
      summary: Search logs
      tags:
        - search
      parameters:
        - name: domain
          in: query
          type: string
        - name: source_ip
          in: query
          type: string
        - name: type
          in: query
          type: string
        - name: class
          in: query
          type: string
        - name: matched
          in: query
          type: boolean
        - name: forwarded
          in: query
          type: boolean
        - name: blocked
          in: query
          type: boolean
        - name: user_id
          in: query
          type: integer
        - name: tags
          in: query
          type: string
        - name: alias
          in: query
          type: string
        - name: date_from
          in: query
          type: string
        - name: time_from
          in: query
          type: string
        - name: date_to
          in: query
          type: string
        - name: time_to
          in: query
          type: string
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          $ref: '#/definitions/SearchResult'
        401:
          $ref: '#/definitions/Response'
        500:
          $ref: '#/definitions/Response'
definitions:
  Response:
    type: object
    properties:
      success:
        type: boolean
        description: Result
      code:
        type: integer
        description: Error code, if any
      message:
        type: string
        description: Short description
      details:
        type: string
        description: Additional data
  Pagination:
    type: object
    properties:
      page:
        type: integer
      pages:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
  Zone:
    type: object
    properties:
      id:
        type: integer
      user_id:
        type: integer
      active:
        type: boolean
      catch_all:
        type: boolean
      forwarding:
        type: boolean
      regex:
        type: boolean
      master:
        type: boolean
        description: For non-admin users a 'master' zone is their unique subdomain. For admins this will always be false.
      domain:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
...
  SearchResult:
    type: object
    properties:
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
      results:
        type: array
        items:
          $ref: '#/definitions/SearchResultItem'
  SearchResultItem:
    type: object
    properties:
      id:
        type: integer
      domain:
        type: string
      source_ip:
        type: string
      type:
        type: string
      cls:
        type: string
      matched:
        type: boolean
      forwarded:
        type: boolean
      blocked:
        type: boolean
      date:
        type: string
...